Commented out some lines that were breaking plugin
A recent update has broken fucntionality for me on two machines in the
last week: user esseguin identified the commented lines as the source.
I've not made any major changes to re implement the check as I'm not
sure what the check was there to achieve, this serves more just to bring
the issue to your attention.

See issue: https://github.com/joonty/vim-phpqa/issues/36