expand file paths
made a docker image with the required tools, and with some aliases etc.
made it possible to run the tools from within a docker container. with
the exception that ${HOME} needs to be mounted in the docker container
and hence when the commands are executed the full path to the file
is needed.

basically changing `@%` to `expand("%:p")`.